Subscription sidebar (#139)
Add a basic sidebar that works on desktop and mobile. It can be in the same sidebar as everything else or in a separate column. The list is limited to 50 entries. If there are more subscriptions than that a "show more" button is added. The list is sorted by the date of the last post in the magazines. That can be switched to a alphabetical sort.