Feature remote retrieval mod followers (#235)
Fetch the remote follower count on updateActor and initial fetch. The remote follower count is used instead of the local follower count if it could be fetched.

The moderators of a magazine are fetched on updateActor and initial fetch. The admin of a local instance will still be the owner of remote magazines and therefore listed in the `moderator` table. There are just too many places in the code that would need to be adjusted, because they rely on the fact that a magazine always has an owner. The owner of remote magazines is therefore hidden in the moderators list.

Implement the `/m/{name}/moderators` endpoint and change the `moderators` property to `attributedTo` in the magazine fetch like other services use (e.g.: Lemmy)