fix duplicate sensitive image state id collision
e-five pointed out that when >1 sensitive posts are right next to each
other that shares the same image, clicking show on either would only
affect the first one.

this is most likely due to the checkbox that stores sensitive state has
an id named after the image id, which would results in html id collision
in such cases where multiple parents shares the same image and this its
id, and the behavior for duplicated html id is effectively undefined.

this adds the the entry/post/comment id that associates with the image
to the part of sensitive state id generation in addition to exising
image id, which should make it harder to collide.