comment collapse: alternate collapse button location
try moving the collapse/expand button to the top right corner of the
comment, after the votes button.

this way the button and the screen stays roughly at the same place when
collapsing/expanding comment,
though this does come with a tradeoff where the button has to be icon
in order to guarantee layout consistency.

not squashing this change yet so it could be switched for testing and
see which one provides better ux in practical uses