add remote url image detection for incoming images
adjust incoming image detection to also try and detect if the given url
leads to an image by downloading and inspecting it with symfony mime
type guesser.

because sometimes a valid url to image doesn't end with image extension
and the path is mostly generated (e.g. uuid filename), in such case
there's no good way to check if it's actually an image besides actually
inspecting the contents

this can be a relatively expensive operation and I can only hope that
most of the url can be detected by preceding checks, and that this check
doesn't get called too often