adjust tag matching to better handle some eastern scripts (#508)
the current hashtags matching and normalizing works fine with latin
scripts, but it often breaks when having to deal with e.g. Japanese
hashtags, which can results in hashtags not being extracted properly or
worse, the posts being dropped entirely

the hashtags matching has beed extended to allow for more characters to
match, and normalization process has been adjusted to better handle
nonlatin script (e.g. Japanese, Thai)

most of the process is a recreation of Mastodon hastags normalization process