Introducing bookmarks In this massive commit bookmarks are added. A user can create multiple bookmarking lists and add entries, posts and comments to them. They are all shown in one view when looking at the list.
One point that is easy to miss: the caching of the twig components `EntryCommentsNestedComponent` and `PostCommentsNestedComponent` was removed (by not using the caching template). It is not really that useful (cache gets cleared as soon as anyone votes on it) and was getting in the way of rendering. That may later be reverted