Add testcase for hashtag matching This commit contains a lot of prep work for testing activity pub code functionally: - create a test only `ApHttpClient` and create an interface for DI - the webfinger code no longer passes a `Request` object in the event, so we can test it easily
The actual test creates 10 remote users and 10 remote magazines with 1 entry each. Then it creates a local magazine that uses the `mbin` hashtag to get posts. Then the 10 remote entries are pulled in each containing the `mbin` hashtag as well, because their magazines are named that way. Afterwards a mastodon style payload is created with an `mbin` hashtag. If all the posts and entries are created correctly then there are exactly 10 announces as the 10 remote users are subscribed to the local magazine