Add instance actor test and remove url field After observing a 401 error from mastodon like this: > {"error":"Verification failed for gehirneimer.de@gehirneimer.de https://gehirneimer.de/i/actor"}
I decided to add a test for the instance actor. I suspect that mastodon actually tried to pull the `url` field of the actor which would result in a 404 causing mastodon to spit out this error. Not 100% sure, but having a pointer to an address we do not serve is pointless, so I removed it